requirements.txt
requests
openai
python-dotenv
flask
pyTelegramBotAPI==4.14.0
